ai_backend/pyproject.toml
[project]
name = "ai-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "qianbeijin",email = "2350105528@qq.com"}
]

requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pinecone-client (>=6.0.0,<7.0.0)",
    "langchain (>=1.2.6,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
    "pypdf (>=6.6.0,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pgvector (>=0.4.2,<0.5.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false